Started implementing the Pack Index format.

+++ src/main/scala/git/Blob.scala
@@ -2,7 +2,7 @@ package git
 
 class Blob extends Object {
   var size = 0
-  var contents: List[Byte] = _ // TODO: Let's not load the data in memory.
+  var contents: List[Short] = _ // TODO: Let's not load the data in memory.
 
   def isBinary: Boolean = {
     // TODO: Try to find at least one null byte.
@@ -11,7 +11,7 @@ class Blob extends Object {
 }
 
 object Blob {
-  def fromObjectFile(bytes: Array[Byte]): Blob = {
+  def fromObjectFile(bytes: Array[Short]): Blob = {
     val o = new Blob
 
     o.contents = bytes.toList
